Changes to demonstrate I. Intent This is to provide a foundation to zooming into higher resolution than Bing (or Google) supports
Motivation.  I want to control, with a map, my rovers to work my yard/garden/sand-worm ranch, and roombas in my house.

II. What is primarily demonstrated
Switching over to a local server (incl. local files) for mapping tiles when the zooming level is high

III. What else is demonstrated
The mouse is hooked in, to allow it to work on the desktop.  This accomplished by creating synthetic touch events

Panner (from the easey project) is used to allow the maps to be moved around.

The browser local storage is used to retrieve/store the center position and zoom (disabled for demo purposes)

The browser geolocation services is used to retrieve the current position if the local storage is not available (disabled for demo purposes)

IV. What was changed?
This is stripped down subset of Mavelous's webapp.